I-7b: one pane surface, not one per engine
What a person does to a rendered browser does not depend on where that
browser runs. A click on an `object-contain` letterbox bar is not a click
on the page whether the frame came from a local Chromium, an Electron
BrowserWindow or a Playwright in a sandbox; a release that drifted onto a
bar has to land in all three, or the page is left holding a button down
forever, mid-selection, with no way for the person to let go.

So the hosted pane does not get a second copy of that. `BrowserPaneSurface`
takes a frame, a holding flag and an `onInput`, and owns the picture, the
pointer arithmetic, the keyboard and the take-control bar. The pointer
helpers underneath it move from `lib/local-browser/client` to
`lib/browser-pane/input` — `toPageCoordinates`, the one-in-flight input
forwarder, the coalescer, the modifier mask — with their tests. The local
client re-exports them under the names its callers already use, so nothing
else moved.

`LocalBrowserBody` keeps what is genuinely local: downloading a Chromium,
minting a frame nonce against device consent, and a lease identity kept in
sessionStorage because there is no signed-in user to be. It is 141 lines
shorter and its suite is untouched and still green, which is the point.

One behaviour is NEW rather than moved: the surface forgets a drag when
the hold ends. A revoked lease cannot send the release — the server would
refuse it — so it is only forgotten, which stops the next pointer move
after taking control again from being clamped onto the page as though the
person were still dragging.
